requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-multipart==0.0.6
jinja2>=3.1.0
Pillow>=10.0.0
qrcode[pil]>=7.4.0
pytesseract>=0.3.10
pyzbar>=0.1.9
reportlab>=4.0.0
httpx>=0.25.0
stripe>=7.0.0
python-dotenv>=1.0.0
pytest>=7.0.0